Add main script and some dependencies
Note: due to the packages being used, the node_modules folder is not in this repository. It contains hundreds of files so to keep this folder less cluttered they were not included. After running the appropriate npm install (package) commands, all dependencies will be correctly loaded and in the right directory.